requirements.txt
streamlit
streamlit_webrtc
av
cv2
numpy
mediapipe
keras.model
webbrowser